[Bronze II] Title: 알파벳 찾기, Time: 0 ms, Memory: 2024 KB -BaekjoonHub

+++ C++20/백준/Bronze/10809. 알파벳 찾기/알파벳 찾기.cc
@@ -18,7 +18,7 @@ int main()
 
 	while (i--)
 	{
-		alphabets[(int)S[i] - 97] = i + 1;
+		alphabets[S[i] - 97] = i + 1;
 	}
 
 	while (i++ < 25)
